Rename old promice dataset file
